+++ packages/gooddata-eval/src/gooddata_eval/core/agentic/kda_skill.py
@@ -185,8 +185,8 @@ class KdaRunResult:
     # Wall-clock time of the turn that called create (None if create never happened) --
     # not any earlier disambiguation turn. See run_agentic_kda_skill's _run_once.
     turn_wall_clock_sec: float | None = None
-    total_turns: float = 0.0
-    total_steps: float = 0.0
+    total_turns: int = 0
+    total_steps: int = 0
     reasoning_steps: list[str] = field(default_factory=list)
     response_id: str | None = None
     tool_call_events: list[ToolCallEvent] = field(default_factory=list)
@@ -279,7 +279,7 @@ def run_agentic_kda_skill(
             all_reasoning_step_events.extend(result.reasoning_step_events or [])
 
         turns = 0
-        steps = 0.0
+        steps = 0
 
         for iteration in range(max_iterations):
             try:
@@ -297,7 +297,7 @@ def run_agentic_kda_skill(
                 turn_completed = False
                 break
             turns += 1
-            steps += float(chat_result.reasoning_step_count)
+            steps += chat_result.reasoning_step_count
             reasoning_steps.extend(chat_result.reasoning_steps or [])
             response_id = chat_result.response_id or response_id
             _accumulate(chat_result)
@@ -337,7 +337,7 @@ def run_agentic_kda_skill(
             actual_create_args=create_args,
             actual_execute_result=execute_result,
             turn_wall_clock_sec=turn_wall_clock_sec,
-            total_turns=float(turns),
+            total_turns=turns,
             total_steps=steps,
             reasoning_steps=reasoning_steps,
             response_id=response_id,
